start splitting quantity methods out into modules
