+++ exoatlas/populations/population_core.py
@@ -1,7 +1,8 @@
 # general class for exoplanet populations
 from ..imports import *
-from ..telescopes import *
-from ..models import *
+
+# from ..telescopes import *
+# from ..models import *
 from .column_descriptions import *
 
 # these are keywords that can be set for a population
@@ -791,742 +792,6 @@ class Population(Talker):
         How many planets are in this population?
         """
         return len(self.standard)
-
-    @property
-    def semimajor_axis(self):
-        """
-        Have a safe way to calculate the semimajor axis of planets,
-        that fills in gaps as necessary. Basic strategy:
-
-            First from table.
-            Then from NVK3L.
-            Then from a/R*.
-
-        """
-
-        # pull out the actual values from the table
-        a = self.standard["semimajoraxis"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(a) == False
-        self.speak(f"{sum(bad)}/{len(self)} semimajoraxes are missing")
-
-        # calculate from the period and the stellar mass
-        P = self.period[bad]
-        M = self.stellar_mass[bad]
-        G = con.G
-        a[bad] = ((G * M * P**2 / 4 / np.pi**2) ** (1 / 3)).to("AU")
-
-        # replace those that are still bad with the a/R*
-        stillbad = np.isfinite(a) == False
-        self.speak(f"{sum(stillbad)}/{len(self)} are still missing after NVK3L")
-        # (pull from table to avoid potential for recursion)
-        try:
-            a_over_rs = self.standard["transit_ar"][stillbad]
-            rs = self.standard["stellar_radius"][stillbad]
-            a[stillbad] = a_over_rs * rs
-        except KeyError:
-            pass
-
-        return a
-
-    @property
-    def semimajor_axis_uncertainty(self):
-        # MASSIVE KLUDGE TO MAKE FLUX ERRORS WORK IN A HURRY!
-        return 0.03 * self.semimajor_axis
-
-    @property
-    def angular_separation(self):
-        """
-        Calculate the angular separation,
-        simply as theta = a/D
-        """
-
-        a = self.semimajor_axis
-        D = self.distance
-
-        theta = np.arctan(a / D).to(u.arcsec)
-
-        return theta
-
-    @property
-    def imaging_contrast(self):
-        """
-        What is the reflected light eclipse depth,
-        for an albedo of 100%?
-
-        But use a kludged radius
-        """
-        return 0.25 * (self.kludge_radius / self.semimajor_axis).decompose() ** 2
-
-    @property
-    def a_over_rs(self):
-        """
-        Have a safe way to calculate the scaled semimajor axis of planets,
-        that fills in gaps as necessary. Basic strategy:
-
-            First from table, mostly derived from transit.
-            Then from the semimajor axis.
-        """
-
-        # pull out the values from the table
-        a_over_rs = self.standard["transit_ar"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(a_over_rs) == False
-        self.speak(f"{sum(bad)}/{len(self)} values for a/R* are missing")
-
-        a = self.semimajor_axis[bad]
-        R = self.stellar_radius[bad]
-        a_over_rs[bad] = a / R
-
-        stillbad = np.isfinite(a_over_rs) == False
-        self.speak(f"{sum(stillbad)}/{len(self)} are still missing after a and R*")
-
-        return a_over_rs
-
-    @property
-    def stellar_luminosity(self):
-        T = self.stellar_teff
-        R = self.stellar_radius
-        sigma = con.sigma_sb
-        return (4 * np.pi * R**2 * sigma * T**4).to(u.Lsun)
-
-    @property
-    def stellar_luminosity_uncertainty(self):
-        T = self.stellar_teff
-        R = self.stellar_radius
-        T_uncertainty = self.get_uncertainty("stellar_teff")
-        R_uncertainty = self.get_uncertainty("stellar_radius")
-
-        sigma = con.sigma_sb
-
-        dLdT = 4 * 4 * np.pi * R**2 * sigma * T**3
-        dLdR = 2 * 4 * np.pi * R**1 * sigma * T**4
-        return (
-            (dLdT**2 * T_uncertainty**2 + dLdR**2 * R_uncertainty**2) ** 0.5
-        ).to("Lsun")
-
-    @property
-    def e(self):
-        """
-        FIXME -- assumes are missing eccentricities are 0!
-        """
-
-        # pull out the actual values from the table
-        e = self.standard["eccentricity"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(e) == False
-        self.speak(f"{sum(bad)}/{len(self)} eccentricities are missing")
-        self.speak(f"assuming they are all zero")
-        e[bad] = 0
-
-        return e
-
-    @property
-    def omega(self):
-        """
-        (FIXME! we need better longitudes of periastron)
-        """
-
-        # pull out the actual values from the table
-        omega = self.standard["omega"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(omega) == False
-        self.speak(f"{sum(bad)}/{len(self)} longitudes of periastron are missing")
-        e_zero = self.e == 0
-        self.speak(f"{sum(e_zero)} have eccentricities assumed to be 0")
-        omega[e_zero] = 0 * u.deg
-
-        return omega
-
-    @property
-    def b(self):
-        """
-        Transit impact parameter.
-        (FIXME! split this into transit and occultation)
-        """
-
-        # pull out the actual values from the table
-        b = self.standard["transit_b"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(b) == False
-        self.speak(f"{sum(bad)}/{len(self)} impact parameters are missing")
-
-        # calculate from the period and the stellar mass
-        a_over_rs = self.a_over_rs[bad]
-        i = self.standard["inclination"][bad]
-        e = self.e[bad]
-        omega = self.omega[bad]
-        b[bad] = a_over_rs * np.cos(i) * ((1 - e**2) / (1 + e * np.sin(omega)))
-
-        # report those that are still bad
-        stillbad = np.isfinite(b) == False
-        self.speak(f"{sum(stillbad)}/{len(self)} are still missing after using i")
-
-        return b
-
-    # the 1360 W/m^2 that Earth receives from the Sun
-    earth_insolation = (1 * u.Lsun / 4 / np.pi / u.AU**2).to(u.W / u.m**2)
-
-    @property
-    def insolation(self):
-        """
-        The insolation the planet receives, in W/m^2.
-        """
-
-        # calculate the average insolation the planet receives
-        insolation = self.stellar_luminosity / 4 / np.pi / self.semimajor_axis**2
-        return insolation.to(u.W / u.m**2)
-
-    @property
-    def insolation_uncertainty(self):
-        """
-        The insolation the planet receives, in W/m^2.
-        """
-
-        # calculate the average insolation the planet receives
-        dinsolation_dluminosity = 1 / 4 / np.pi / self.semimajor_axis**2
-        dinsolation_dsemimajor = (
-            -2 * self.stellar_luminosity / 4 / np.pi / self.semimajor_axis**3
-        )
-        L_uncertainty = self.get_uncertainty("stellar_luminosity")
-        a_uncertainty = self.get_uncertainty("semimajor_axis")
-        insolation_uncertainty = (
-            dinsolation_dluminosity**2 * L_uncertainty**2
-            + dinsolation_dsemimajor**2 * a_uncertainty**2
-        ) ** 0.5
-        return insolation_uncertainty.to(u.W / u.m**2)
-
-    @property
-    def relative_insolation(self):
-        """
-        The insolation the planet receives, relative to Earth.
-        """
-        return self.insolation / self.earth_insolation
-
-    @property
-    def relative_insolation_uncertainty(self):
-        """
-        The insolation the planet receives, relative to Earth.
-        """
-        return self.insolation_uncertainty / self.earth_insolation
-
-    @property
-    def log_relative_insolation(self):
-        return np.log10(self.relative_insolation)
-
-    @property
-    def relative_cumulative_xuv_insolation(self):
-        """
-        A *very very very* approximate estimate for the cumulative XUV flux (J/m**2)
-        felt by a planet over its lifetime. It comes from Zahnle + Catling (2017) Equation 27,
-        where they say they did integrals over the Lammer et al. (2009) XUV flux relations.
-        This effectively assumes that the early times dominate, so the time integral doesn't
-        depend (?!?) on the age of the system. It's very rough!
-        """
-        xuv_proxy = (self.stellar_luminosity / u.Lsun) ** -0.6
-        return self.relative_insolation * xuv_proxy
-
-    @property
-    def teq(self):
-        """
-        The equilibrium temperature of the planet.
-        """
-        f = self.insolation
-        sigma = con.sigma_sb
-        A = 1
-        return ((f * A / 4 / sigma) ** (1 / 4)).to(u.K)
-
-    @property
-    def planet_luminosity(self):
-        """
-        The bolometric luminosity of the planet (assuming zero albedo).
-        """
-        return (self.teq**4 * con.sigma_sb * 4 * np.pi * self.radius**2).to(u.W)
-
-    @property
-    def transit_depth(self):
-        """
-        The depth of the transit
-        (FIXME, clarify if this is 1.5-3.5 or what)
-        """
-
-        # pull out the actual values from the table
-        d = self.standard["transit_depth"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(d) == False
-        self.speak(f"{sum(bad)}/{len(self)} transit depths are missing")
-
-        Rp = self.radius[bad]
-        Rs = self.stellar_radius[bad]
-
-        d[bad] = (Rp / Rs).decompose() ** 2
-
-        # report those that are still bad
-        stillbad = np.isfinite(d) == False
-        self.speak(f"{sum(stillbad)}/{len(self)} are still missing after Rp/Rs")
-
-        return d
-
-    @property
-    def transit_duration(self):
-        """
-        The duration of the transit
-        (FIXME, clarify if this is 1.5-3.5 or what)
-        """
-
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-
-            # pull out the actual values from the table
-            d = self.standard["transit_duration"].copy()
-
-            # try to replace bad ones with NVK3L
-            bad = np.isfinite(d) == False
-            self.speak(f"{sum(bad)}/{len(self)} transit durations are missing")
-
-            P = self.period[bad]
-            a_over_rs = self.a_over_rs[bad]
-            b = self.impact_parameter[bad]
-
-            T0 = P / np.pi / a_over_rs
-            T = T0 * np.sqrt(1 - b**2)
-
-            e = self.e[bad]
-            omega = self.omega[bad]
-            factor = np.sqrt(1 - e**2) / (1 + e * np.sin(omega))
-
-            d[bad] = (T * factor).to(u.day)
-
-            # report those that are still bad
-            stillbad = np.isfinite(d) == False
-            self.speak(
-                f"{sum(stillbad)}/{len(self)} are still missing after P, a/R*, b"
-            )
-
-            return d
-
-    @property
-    def kludge_mass(self):
-        """
-        Have a safe way to calculate the mass of planets,
-        that fills in gaps as necessary. Basic strategy:
-
-            First from table.
-            Then from msini.
-        """
-
-        # pull out the actual values from the table
-        M = self.standard["mass"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(M) == False
-        self.speak(f"{sum(bad)}/{len(self)} masses are missing")
-
-        # estimate from the msini
-        try:
-            M[bad] = self.msini[bad]
-        except (KeyError, AssertionError, AtlasError, AttributeError):
-            pass
-
-        # replace those that are still bad with the a/R*
-        stillbad = np.isfinite(M) == False
-        self.speak(f"{sum(stillbad)}/{len(self)} are still missing after msini")
-
-        return M
-
-    @property
-    def kludge_radius(self):
-        """
-        Have a safe way to calculate the radii of planets,
-        that fills in gaps as necessary. Basic strategy:
-
-            First from table.
-            Then from mass, via Chen & Kipping (2017).
-        """
-
-        # pull out the actual values from the table
-        R = self.standard["radius"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(R) == False
-        self.speak(f"{sum(bad)}/{len(self)} radii are missing")
-
-        # estimate from Chen and Kipping
-        try:
-            M = self.kludge_mass
-            R[bad] = estimate_radius(M[bad])
-        except (KeyError, AssertionError, AtlasError, AttributeError):
-            pass
-
-        # replace those that are still bad with the a/R*
-        stillbad = np.isfinite(R) == False
-        self.speak(
-            f"{sum(stillbad)}/{len(self)} are still missing after Chen & Kipping (2017)"
-        )
-
-        return R
-
-    @property
-    def kludge_age(self):
-        """
-        Have a safe way to calculate the age of planets,
-        that fills in gaps as necessary. Basic strategy:
-
-            First from table.
-            Then assume 5 Gyr.
-        """
-
-        # pull out the actual values from the table
-        age = self.standard["stellar_age"].copy()
-
-        # try to replace bad ones with NVK3L
-        bad = np.isfinite(age) == False
-        self.speak(f"{sum(bad)}/{len(self)} ages are missing")
-
-        # estimate from the msini
-        try:
-            age[bad] = 5 * u.Gyr
-        except (KeyError, AssertionError, AtlasError, AttributeError):
-            pass
-
-        # replace those that are still bad with the a/R*
-        stillbad = np.isfinite(age) == False
-        self.speak(
-            f"{sum(stillbad)}/{len(self)} are still missing after blindly assuming 5Gyr for missing ages"
-        )
-
-        return age
-
-    @property
-    def surface_gravity(self):
-        """
-        (FIXME) -- make an assumption for planets without masses
-        """
-
-        G = con.G
-        M = self.mass
-        R = self.radius
-
-        g = (G * M / R**2).to("m/s**2")
-        return g
-
-    @property
-    def density(self):
-        """
-        The density of the planet.
-        """
-        mass = self.mass
-        volume = 4 / 3 * np.pi * (self.radius) ** 3
-        return (mass / volume).to("g/cm**3")
-
-    @property
-    def escape_velocity(self):
-        """
-        The escape velocity of the planet.
-        """
-        G = con.G
-        M = self.mass
-        R = self.radius
-        return np.sqrt(2 * G * M / R).to("km/s")
-
-    @property
-    def orbital_velocity(self):
-        return (2 * np.pi * self.semimajor_axis / self.period).to("km/s")
-
-    @property
-    def impact_velocity(self):
-        return np.sqrt(self.orbital_velocity**2 + self.escape_velocity**2)
-
-    @property
-    def escape_parameter(self):
-        """
-        The Jeans atmospheric escape parameter for atomic hydrogen,
-        at the equilibrium temperature of the planet.
-        """
-        k = con.k_B
-        T = self.teq
-        mu = 1
-        m_p = con.m_p
-        G = con.G
-        M = self.mass
-        R = self.radius
-
-        e_thermal = k * T
-        e_grav = G * M * m_p / R
-        return (e_grav / e_thermal).decompose()
-
-    @property
-    def distance_modulus(self):
-        """
-        The distance modulus to the system, in magnitudes.
-        """
-        mu = 5 * np.log10(self.distance / (10 * u.pc))
-        return mu
-
-    def scale_height(self, mu=2.32):
-        """
-        The scale height of the atmosphere, at equilibrium temperature.
-        """
-        k = con.k_B
-        T = self.teq
-        m_p = con.m_p
-        g = self.surface_gravity
-        return (k * T / mu / m_p / g).to("km")
-
-    def transmission_signal(self, mu=2.32, threshold=2):
-        """
-        What is the transit depth of 1 scale height of an
-        atmosphere transiting in front of the star.
-
-        Parameters
-        ----------
-        mu : float
-            Mean molecular weight (default 2.2 for H/He)
-        threshold : float
-            By how many sigma must the planet mass be detected?
-
-        """
-        with np.errstate(invalid="ignore"):
-            H = self.scale_height(mu)
-            Rp = self.radius
-            Rs = self.stellar_radius
-            depth = (2 * H * Rp / Rs**2).decompose()
-
-            dlnm = self.get_uncertainty("mass") / self.mass
-            bad = dlnm > 1 / threshold
-            depth[bad] = np.nan
-            return depth
-
-    def reflection_signal(self, albedo=0.1):
-        """
-        What is the reflected light eclipse depth,
-        for an albedo of 100%?
-        """
-        return albedo * 0.25 * (self.radius / self.semimajor_axis).decompose() ** 2
-
-    def emission_signal(self, wavelength=5 * u.micron):
-        """
-        What is the thermal emission eclipse depth,
-        assuming Planck spectra for both star and planet?
-
-        This calculation assumes a Bond albedo of 0
-        and that heat is uniformly distributed over the planet.
-
-        Parameters
-        ----------
-        wavelength : astropy.unit.Quantity
-            The wavelength at which it should be calculated.
-        """
-
-        # create thermal emission sources for both star and planet
-        import rainbowconnection as rc
-
-        star = rc.Thermal(teff=self.stellar_teff, radius=self.stellar_radius)
-        planet = rc.Thermal(teff=self.teq, radius=self.radius)
-
-        # calculate the depth as the luminosity ratio
-        depths = planet.spectrum(wavelength) / star.spectrum(wavelength)
-
-        return depths
-
-    def stellar_brightness(self, wavelength=5 * u.micron):
-        """
-        How many photons/s/m^2/micron do we receive from the star?
-
-        This is calculated from the distance, radius, and
-        stellar effective temperature of the stars.
-
-        (It could be potentially be improved with PHOENIX
-        model grids and/or cleverness with photometry.)
-
-        Parameters
-        ----------
-        wavelength : astropy.unit.Quantity
-            The wavelength at which it should be calculated.
-        """
-
-        # import some tools for easy cartoon spectra
-        import rainbowconnection as rc
-
-        # create source with right temperature, size, distance
-        teff, radius = self.stellar_teff, self.stellar_radius
-        star = rc.Thermal(teff=teff, radius=radius).at(self.distance)
-
-        # calculate the energy flux
-        flux_in_energy = star.spectrum(wavelength)
-
-        # convert to photon flux
-        photon_energy = con.h * con.c / wavelength / u.ph
-        flux_in_photons = flux_in_energy / photon_energy
-
-        # return the
-        return flux_in_photons.to("ph s^-1 m^-2 micron^-1")
-
-    def stellar_brightness_in_telescope_units(self, telescope_name="JWST", **kw):
-        """
-        The stellar brightness, converted to telescope units.
-
-        Parameters
-        ----------
-        telescope_name : str
-            The name of the telescope.
-
-        wavelength : astropy.unit.Quantity
-            The wavelength at which it should be calculated.
-
-        R : float
-            The spectral resolution at which the
-            telescope will bin wavelengths.
-
-        dt : astropy.units.quantity.Quantity
-            The time over which the telescope exposes.
-        """
-
-        # what counts as 1 "telescope unit" (e.g. JWST at R=20 at 5 microns for 1 hour)
-        telescope_unit = define_telescope_unit_by_name(telescope_name, **kw)
-
-        # what's the photon flux (photons/m**2/s)
-        flux_in_photons = self.stellar_brightness(telescope_unit.wavelength)
-
-        # quote the brightness as (for example) gigaphotons/JWST at R=20 at 5 microns in 1 hour
-        unit = lotsofphotons_unit / telescope_unit
-        return flux_in_photons.to(unit)
-
-    def depth_uncertainty(
-        self, telescope_name="JWST", per_transit=False, dt=1 * u.hour, **kw
-    ):
-        """
-        What is the transit/eclipse depth uncertainty
-        with a particular telescope
-        at a particular wavelength
-        at a particular resolution?
-
-        By default, this will be calculated for one transit.
-        Optionally, it can be calculated for a given amount of time instead.
-
-        Parameters
-        ----------
-        telescope_name : str
-            The name of the telescope.
-
-        per_transit : bool
-            If True, calculate the depth uncertainty for one transit.
-            If False, calculate the depth uncertainty for a certain amount
-            of in-transit time. You likely want to specify `dt` as a
-            keyword argument to set that amount of in-transit time.
-            In either case, an out-of-transit baseline equal to the
-            total in-transit time will be assumed. This means the actual
-            time cost will be twice the transit duration or `dt` chosen,
-            and the depth uncertainty will be a factor sqrt(2) larger
-            than the pure photon noise binned to the relevant timescale.
-
-        wavelength : astropy.unit.Quantity
-            The wavelength at which it should be calculated.
-
-        R : float
-            The spectral resolution at which the
-            telescope will bin wavelengths.
-
-        dt : astropy.units.quantity.Quantity
-            The time over which the telescope exposes. If `per_transit=True`,
-            this will be ignored. Otherwise, it will set the total amount
-            of in-transit time observed, assuming that an equal amount of
-            time will *also* be observed out of transit.
-        """
-
-        # what counts as 1 "telescope unit" (e.g. JWST at R=20 at 5 microns for 1 hour)
-        telescope_unit = define_telescope_unit_by_name(telescope_name, dt=dt, **kw)
-
-        # what's the photon flux (photons/m**2/s)
-        flux_in_photons = self.stellar_brightness(telescope_unit.wavelength)
-
-        # what's the total collecting power?
-        if per_transit:
-            ratio_of_collecting_time = self.transit_duration / dt
-        else:
-            ratio_of_collecting_time = 1
-        collecting_power = 1 * telescope_unit * ratio_of_collecting_time
-
-        # what's the total number of photons collected during transit
-        N = (flux_in_photons * collecting_power).to(u.ph).value
-
-        # what's the flux uncertainty on the time scale of one transit?
-        sigma = 1 / np.sqrt(N)
-
-        # inflate by a factor of sqrt(2) for equal out-of-transit
-        oot = np.sqrt(2)
-        sigma_depth = sigma * oot
-
-        return sigma_depth
-
-    def _get_noise_and_unit(self, telescope_name="JWST", per_transit=False, **kw):
-        """
-        Tiny helper to get the noise and the telescope_unit
-        for a telescope observation of a planet.
-        """
-
-        # figure out the noise
-        noise = self.depth_uncertainty(
-            telescope_name=telescope_name, per_transit=per_transit, **kw
-        )
-
-        # create a telescope unit (mostly to get a default wavelength)
-        telescope_unit = define_telescope_unit_by_name(telescope_name, **kw)
-
-        return noise, telescope_unit
-
-    def depth_snr(self, telescope_name="JWST", **kw):
-        """
-        What's the approximate S/N for the detection of the planet's transit?
-        """
-
-        noise, telescope_unit = self._get_noise_and_unit(
-            telescope_name=telescope_name, **kw
-        )
-        signal = self.transit_depth
-        return signal / noise
-
-    def emission_snr(self, telescope_name="JWST", **kw):
-        """
-        What's the approximate S/N for the detection of the
-        thermal emission eclipse of a planet?
-        """
-
-        noise, telescope_unit = self._get_noise_and_unit(
-            telescope_name=telescope_name, **kw
-        )
-        signal = self.emission_signal(wavelength=telescope_unit.wavelength)
-        return signal / noise
-
-    def reflection_snr(self, telescope_name="JWST", albedo=1, **kw):
-        """
-        What's the approximate S/N for the detection of the
-        reflected light eclipse of a planet?
-        """
-
-        noise, telescope_unit = self._get_noise_and_unit(
-            telescope_name=telescope_name, **kw
-        )
-        signal = self.reflection_signal(albedo=albedo)
-        return signal / noise
-
-    def transmission_snr(self, telescope_name="JWST", mu=2.32, threshold=2, **kw):
-        """
-        What's the approximate S/N for the detection of the
-        reflected light eclipse of a planet?
-        """
-
-        noise, telescope_unit = self._get_noise_and_unit(
-            telescope_name=telescope_name, **kw
-        )
-        signal = self.transmission_signal(mu=mu, threshold=threshold)
-        return signal / noise
 
     def scatter(
         self, xname, yname, c=None, s=None, names=True, xlog=True, ylog=True, **kw
